dutch-speech-to-text-test/pyproject.toml
[project]
name = "dutch-speech-to-text"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-speech>=2.33.0",
    "google-cloud-storage>=3.4.0",
    "ipykernel>=6.30.1",
]
